fix: moved hook to SearchPage.component.js and renamed files
