+++ src/core_modules/capture-core/components/Pages/Search/hooks/buildSearchGroup.js
@@ -1,11 +1,11 @@
 // @flow
-import { getUserStorageController } from '../../storageControllers';
-import { userStores } from '../../storageControllers/stores';
-import { SearchGroupFactory } from '../../metaDataMemoryStoreBuilders/common/factory';
+import { getUserStorageController } from '../../../../storageControllers';
+import { userStores } from '../../../../storageControllers/stores';
+import { SearchGroupFactory } from '../../../../metaDataMemoryStoreBuilders/common/factory';
 import type {
     ProgramTrackedEntityAttribute,
     TrackedEntityAttribute,
-} from '../../components/WidgetProfile/DataEntry/FormFoundation/types';
+} from '../../../WidgetProfile/DataEntry/FormFoundation/types';
 
 type SearchProgramOrTET = {|
     minAttributesRequiredToSearch: number,
