+++ src/core_modules/capture-core/components/Pages/Search/hooks/index.js
@@ -1,2 +1,3 @@
 
 export { buildSearchGroup } from './buildSearchGroup';
+export { useSearchOption } from './useSearchOption';
